+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-03 15:40:44
+
+### Changed Files
+- `src/features/targets/TargetsPage.tsx`
+- `src/features/targets/components/WelcomeTargetCard.tsx`
+
+### Statistics
+```
+ 2 files changed, 60 insertions(+), 22 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-03 15:31:07
 
 ### Changed Files